build-git: git complains when master is specified twice, fix that
In cc0f714f293f ("build-git: Make it possible to specify branch as $1"),
I made it possible to compile a special branch or tag if it was
specified as $1. I knew that this would execute commands like

  git push -f REMOTE master master html maint man next pu todo

but I thought it would be OK. But I got this message from the grumpy old
git if $1 is not specified:

  error: dst ref refs/heads/master receives from more than one src.

So, in push_changes(), define a temporary $push_local variable to take
care of that.

3f8db306-0311-11e1-8ef1-95f0de5972ee