build-git: Dont't push commits after the build, only tags
Sometimes I forget about the build, have it lying around in some
screen(1) session somewhere, and if I've run another build on another
machine with newer commits in the meantime, the remote branches would be
reset to the old version. It pushes the commits in the beginning of the
build anyway, so it isn't necessary to have another one at the end.

b0146a1c-6567-11e1-a040-8000600fe800