build-git: Push branches locally instead of delete+recreate
Instead of deleting and then recreate the branches, push the remote
branches to the local repo. A bonus feature by this method is that we
can check that the master branch fast-forwards properly. If it doesn't,
there's some fishy things going on, as master never should be rebased.
At least that's what they say.

e856af3a-069c-11e2-aae3-fefdb24f8e10