Makefile: remotes: google: Add Google Code
Just an experiment and I might change my mind at some point. There's
something in the "Export Control" additional terms at
<https://code.google.com/projecthosting/terms.html> I'm not too fond of:

  Users residing in countries on the United States Office of Foreign
  Assets Control sanction list, including Cuba, Iran, North Korea, Sudan
  and Syria, may not post or access Content available through the Google
  Code website.

FOSS software should be available to anyone, independent of politics or
which country they're born in. But it's just a mirror anyway, and Google
is awesome. Not their fault that they have to obey those laws.

089be4f0-1b19-11e2-b13f-fefdb24f8e10