suuid.t: Create $cdata and tighten up a --count test
The "Log contents OK after count, comment and tag" was missing a {5} in
the regexp, but did succeed even when it shouldn't. The reason for this
is that the regexp was greedy and matched everything from '<host>.+?'
until the end. Fix that by replacing '.+?' with $cdata, which is defined
as '[^<]+'. The $cdata variable is used because all the other tests will
use it in the next commit.

db37128c-3c42-11e2-97ad-b627d546bf76