FLDBsum.pm: Digest::SHA1 is renamed to Digest::SHA upstream
Use "Digest::SHA" instead of "Digest::SHA1" and add "(1)" to the
declaration of $sha1 to indicate that we want a SHA1. It's not strictly
necessary, as SHA1 is the default hash to use, but that might change in
the future.

I hope this doesn't lead to incompatibility in other distros, but it
seems as this package (libdigest-sha-perl) has been around for some
time, so it should probably be available. And if it's not, the module is
available from CPAN.

693416b0-be17-11e2-a0ff-001f3b596ec9