fldb: process_file: Add "-l" check and ignore the file if it's a symlink
It's not enough to check that the current file actually is a file by
using "-f". If the symlink points to an existing file, -f still returns
true. Add an additional test using "-l" to check if it's a symlink. If
it is, ignore it.

f267c81c-be69-11e2-ae40-4262fb1f76e6