Makefile: obsolete: New rule for deleting old refs everywhere
Some deleted branches have kept reappering everywhere because I run
things like "git update-dirs -rA" in ~/src/git/ on a random computer. If
there's a repo in that subtree that has any of those defined, it comes
back and spreads to all other available repos. It doesn't even have to
be a local branch, it's enough that one of the remotes in that repo
contains the branch. And that's totally awesome. But it's a bit tiresome
to delete the obsolete branches again and again and again and again and
again and again and again og sånn går nu dagan.

2c1b12cc-ead3-11e2-a2d7-0800274013ad
a5dc30be-ead3-11e2-b0dd-0800274013ad