git-dangling: Store dangling blobs too
At the moment it just stores them as "blob-[SHA1]" files in the current
directory. Should probably have some standard directory where they are
stored, but I'll have to think about that. Should also create a function
of this, there's some code duplication. Nevermind for now. A FIXME for
later.

988b838e-9b62-11e3-b5f6-001f3b596ec9