Revert "build-git: Don't install as root"
This reverts commit 0620380497373860c174952555603e1ebd1794e5. It's kind
of problematic that it wants to create the directory beforehand;
firstly, it's annoying that we have to enter the root password twice,
and secondly, if I'll run a test compile of for example 'next', it wants
to create an empty directory there even if we're not going to use it.
Probably an easy fix, but I don't have time to look into it now.

189d2dfa-d463-11e3-a9c3-fefdb24f8e10