git-dangling: Don't delete $dangfile if it's inside the .git dir
As mentioned in the previous commit, storing $dangfile in the .git
directory should be jolly good. But if we're in a bare repo, there's no
.git directory to store it in, so in that case it's still deleted.
Another thing is that "git rev-parse --show-toplevel" doesn't work in
bare repos or inside the .git directory, so it should take care of that
too.

d35ee4b4-2fb5-11e4-b3f4-c80aa9e67bbd