installed_progs.t: installed(): Add argument for stderr/stdout/both
Before this, stderr was redirected to stdout and both streams were
checked at the same time. Some programs insist on being difficult,
though, for example sending lots of rubbish to stderr while the version
info (which is not very informative) is sent to stdout. Because I'm so
damn polite I won't write the name of the program that made me do this.
It's not like I'll add it in the next commit or anything.

6be3dad6-47e5-11e4-9c58-c80aa9e67bbd