mc: Add -u to disable subshell support AGAIN
The problem this time is that output that should go to the terminal is
randomly swallowed somewhere (at least with version 4.8.13, compiled
under Linux Mint 17, and probably the original version too), for example
when sending text to stdout with cat(1). Subshell support is nice, but
this is unacceptable behaviour. A workaround is to just start bash(1)
from inside mc. Will probably not enable subshell support again, it's
been fucking up stuff since 2004 or something.

c9bbad42-dccc-11e4-b16b-000df06acc56