ga-getnew: Add --no-run-if-empty to xargs
If there were no files listed (for example if there wasn't any activity
the last month), it would run ga get and ga fsck on the whole
repository, and that's what was to be avoided in the first place.

Using the long version of the option instead of -r because it's a GNU
extension.

eabf33b4-f492-11e4-8fa1-000df06acc56