tohex: Disallow command line arguments, only read from stdin
It seems like a magnificent project to get UTF-8 to work properly on
both stdin and filehandles. There are some "use" things I could throw at
it, but then it refuses to read text not in UTF-8. For now, the easiest
solution is to only use stdin, but that's the way I mostly use those
things anyway.

0b24b252-2b55-11e5-bd5b-000df06acc56