ustr: Add "use utf8" to fix strikethrough example in usage() + test
I stopped using "use utf8" in Perl source in 2003, it had a tendency to
mess up more than it fixed, but in this case it was necessary. Maybe
I'll give it another chance. Things surely must have gotten better the
last 12 years. For example, in the log message for commit 1ebe4fa
(2004-01-07 03:51:27 +0100) I ran a comparison with and without "use
utf8". The script executed 10 times faster without "use utf8". That's
fixed now, there's no significant difference with or without. In fact,
it seems as it's a tiny bit faster with "use utf8", which is a nice
surprise.

09841ee4-3792-11e5-b328-000df06acc56