build-vim: Check that $srcdir is clean earlier and only if not --dry-run
Check that $srcdir is clean before creating any directories. Also
perform this test only if no simulation is done. This is in case the
directory doesn't exist, it would abort in that case if --dry-run was
used. The whole point of -n/--dry-run is to show the whole compilation
process, it will bail out at this point when it's run for real.

7376bb9a-67fd-11e5-af0b-efc55bf34391