git-update-dirs: Make "--all-options -a" actually call "git allbr -a"
"--all-options -a" didn't execute "git allbr -a" because $Opt{'allbr'}
never reached 2 which is the value it needs to run "git allbr -a" in
non-bare repositories. To make it work, it was necessary to specify
"-Aaa".

Use "$Opt{'allbr'} += 1" instead of "$Opt{'allbr'} = 1" to fix it.

c0193db8-6c19-11e5-b557-fefdb24f8e10