edit-sqlite3.t: Modify test to check for SQL instead of binary diff
The "Backup file is identical to ok.sqlite" test checked if sqlite3
generated a new file that was identical to the old one. This is almost
certainly doomed to fail every time a new SQLite version is released, so
only check that the SQL is similar. Change description to "Data of
backup file is identical to ok.sqlite".

b41bbd00-7395-11e5-84ae-fefdb24f8e10