Store contents from synced.sqlite as SQL in synced.sql instead
Having the file sync info in an in-repo SQLite database works fine, but
it's merge hell. Executing "std" commands or anything else that modifies
synced.sqlite on branches while there are simultaneous changes on master
almost always results in conflicts and I'll have to dump to text files,
edit and then import it back. Urgh. So instead, store the info in a flat
SQL file which is simply a "sort-sqlite synced.sqlite; sqlite3 .dump
>synced.sql". Makes life about 7.52% easier overall.

This commit does the following:

- "sqlite3 synced.sqlite .dump >synced.sql"
- "git ignore synced.sqlite", this adds synced.sqlite to .gitignore and
  deletes it from Git.
- Add "sqlite3 synced.sqlite .dump >synced.sql" to Makefile so it's
  executed with "make". Should be done before commit when there are new
  filesync changes.

15842b22-7b02-11e5-91ef-fefdb24f8e10