std.t: Add chdir to make some tests work in other repositories
std(1) changes to ~/bin/ to read commit SHA-1 of current HEAD, and that
might differ from the SHA-1 in the current repository. Change directory
to ~/bin/ temporarily to get the SHA-1 of HEAD there, and use that in
the tests. This whole thing with changing to ~/bin/ is sort of a FIXME,
and I've marked it as such. Will probably add some configuration option
for this sometime.

2d682a76-9183-11e5-adef-fefdb24f8e10