build-postgis: Remove "git checkout -f ./install-sh"
This was a file that always got modified after compiling, so it had to
be reset. It was removed from the "svn-2.2" branch in
git@github.com:postgis/postgis.git in commit 7c3d25a ("install-sh is a
built artifact, should not be in the repository", 2014-07-01) where they
said... Well, that.

382be10e-b349-11e5-859c-02010e0a6634