latex-article: Update it to the 21th century, use UTF-8
LilyPond has brought back my love for LaTeX. I used it a lot in the old
days, but there were some issues, for example lack of proper Unicode
support and small annoying things like not being able to use $Id$ in the
header. I don't use that nowadays, and the Unicode support is probably
OK these days, so let's have a look at it again. Haven't made any
LaTeX-related changes to it in 17 years, way back in 1999. o.O

41104ddc-f949-11e5-ac92-02010e0a6634