mkt: $GZIP_CMD: Add --rsyncable
The file gets a little bigger, but it's really worth it. Works like a
charm. Now it's official too, it was added to gzip.git in commit d6d48f4
("gzip: support the --rsyncable option", 2016-03-02), 7 months ago. It's
also known as v1.6-42-gd6d48f4 among friends (git describe). Before
that, it was just an unofficial patch floating around on the web
somewhere. More info here:

  http://lists.gnu.org/archive/html/bug-gzip/2012-06/msg00023.html
  http://superuser.com/questions/653057/caveats-with-gzip-rsyncable-no-speedup
  https://bugs.launchpad.net/ubuntu/+source/gzip/+bug/1021975

4b722152-2b6b-11e6-9cac-fefdb24f8e10