sort-sqlite: Make it work with multi-line rows, add -c/--column
The first version was an ugly hack (it even had a warning about it) that
parsed the dump file from sqlite3 and then regenerated the database. It
worked well, and I've never experienced problems with it. But do it more
properly, copy all rows to a temporary table, delete all rows from the
original table, and copy them back in sorted order.

de49d85c-a23b-11e6-b1ca-e2363df49851