Rename ts to cts, "Create Task and Start it"
All these Taskwarrior/Timewarrior scripts have pretty similar names, so
make it a bit easier to remember. To sum it up:

t   - Start task(1) (Taskwarrior) and automatically commit changes in
      the Taskwarrior database to Git.
cts - Create a new task in Taskwarrior and mark it as started.
tw  - Start timew(1) (Timewarrior), behave the same way as t.
tws - Start a new session in Timewarrior and use all the arguments in a
      single description.

6c589028-1e62-11e7-ab5e-db5caa6d21d3