log_df: Create "meta" table with dbversion and create upgrade_db()
There'll probably be some changes to the database schema later, create a
"meta" table where the current layout version number is stored so it's
able to automatically upgrade to new versions. At the moment it creates
the meta table and sets dbversion to 1 if it doesn't exist. It creates a
backup of the database in the log directory before upgrade, so things
should be safe.

c9774f00-9893-11e7-b86f-f74d993421b0