list-youtube: Make it work with filenames containing '|' (U+007C)
There's no way to disable the separator when using the .import command
in sqlite3. Any '|' (the default separator character) in filenames is
interpreted as a separator, resulting in an "expected 1 columns but
found 2 - extras ignored" warning and the line is truncated.

To fix this, create a mkinsert() function that converts the input to
INSERT statements and use it before the sqlite3 command.

13120e14-6892-11e9-8608-4f45262dc9b5