ga-infounused: Use `--no-textconv` in `git diff`
In case there's a git textconv filter in the repo that doesn't work with
the current setup. For example, if there's a repo that uses the
`rdbl-midi` filter and Lilypond isn't installed, it generates a
"lilymidi: command not found" error message.

That seems oddly specific, you may think. Well, that's what happened.

6d2f4f1c-6ad9-11ee-af68-754a1be59913