mdisk: Avoid delay when mounting disks with a `/.dontsleep/` directory
Some disks experience a delay before the first write, possibly due to
journal preparation. The duration of this delay can vary; in this
instance, it was 33 seconds. This commit eliminates the delay by
combining the creation and execution of the script into a single
background process.

e6dd6ae0-9678-11ee-8d3d-83850402c3ce