bpakk: Output terminating `\n` at the end
It turns out it's not necessary to drop the final `\n` when using
`:'<,'>!bpakk` in Vim. I thought it would insert an empty line under the
visually marked area. Good thing, because it was pretty annoying not to
have the `\n` at the end when using it from the command line.

6bf7c876-7c63-11ef-9c9d-83850402c3ce