ga-fsck-size.t: Strip fsck lines, git-annex processes files in random order
No matter what I do, it seems to be random. Had to do something like

```sh
find -type l -print0 | LC_COLLATE=C sort -z \
| xargs -0 -L 1 git annex fsck 2>&1 | ../../$CMD
```

to make it work in a predictable way, and that's not very elegant. Oh
well, the "fsck" lines aren't important; the important part is the
output `ga-fsck-size` creates and that it counts correctly.

3d8c91d0-8427-11ef-b23d-83850402c3ce