c/src/Makefile: test: Don't use tests in `t/`, use --selftest, 73:11
The tests in src/t/ are soon obsolete, but are still used by `make
valgrind` until --valgrind is in place.

For the next commits while this is going on, the output from `make tlok`
is stored in the commit subject line. The format is "a:b" where `a` is
the number of untested lines and `b` is the number of tests at the time
of commit.

b27e0ab0-ad93-11ef-a7bc-83850402c3ce