std/c/: `opt` isn't global anymore, declare it as `static`
Create `const struct Options *` parameter for all functions that need
access to the members of `opt`. This access is read-only, but for local
modifications, a copy of `opt` can be stored in a local variable with
opt_struct(). An example of this is in test_valgrind_option() where the
command that checks if Valgrind is installed is executed without
Valgrind.

e4250022-3822-11f0-b2b9-83850402c3ce