selftest.c: Use EXECSTR in tests containing `argv[0]` output, 0:119
As described in previous commit `48ec65d4d2cb` ("test_command(): Replace
text in stdout and stderr with `argv[0]` value"), test_command() now
replaces `__EXSTR__` in stdout and stderr with the value of `argv[0]`.
This enables the use of tc() (full stderr and stdout verification)
instead of just substring verification with sc() in those tests.

76ea3066-55f7-11f0-afad-83850402c3ce