selftest.c: Print all "# " lines to stdout instead of stderr
It emulated the behavior from Test::More in Perl, sending "ok"/"not ok"
lines to stdout and "# " lines to stderr. This made it necessary to
redirect stderr to stdout in the shell to show all the "# " lines in a
viewer like `less`(1). "# " is part of the TAP standard, so it's
expected that TAP analyzers ignore "# " lines, or only process "ok"/"not
ok" lines. Sending everything to stdout also seems to be the common
behavior among other testing frameworks.

Commit ID: 736121ee-79dc-11f0-b0f3-83850402c3ce