selftest.c: Add OK_MEMCMP and OK_MEMCMP_L macros and break long lines
The macros have the same arguments as OK_STRNCMP and OK_STRNCMP_L, but
the buffers can contain null bytes.

The line got a bit long when adding it to test_ok_macros(), so it had to
be wrapped. And when that line is wrapped, all the other long lines in
the function might as well be wrapped.

Commit ID: f1fda0e8-7e7a-11f0-87d2-83850402c3ce