std/c/: Make it possible to chdir() during testing, 0:173
Create get_execname() that returns an absolute path to `execname` if
necessary so the executable can change directories during testing, for
example to a temporary directory. This sometimes requires `progname` to
be set to the same value as `execname` when functions are tested, so
make `progname` global.

Commit ID: 29db3d22-9a5f-11f0-9337-83850402c3ce